+++ src/App.jsx
@@ -1,19 +1,15 @@
 
 import './App.css'
-
+import TodoList from './TodoList.jsx';
+import TodoForm from './TodoForm.jsx'
 function App() {
-  const todos = [
-    {id: 1, title: "Review Week 1"},
-    {id: 2, title: "Complete Week 1 Assignment"},
-    {id: 3, title: "Submit Assignment"}
 
-  ]
   return (
     <div>
       <h1>My Todos</h1>
-      <ul>
-        {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
-      </ul>
+      <TodoForm />
+      <TodoList />
+      
     </div>
   )
 }
